[build] fix gendarme download
Github seems to have changed their download URLs, our old URL is no
longer working:
https://cloud.github.com/downloads/spouliot/gendarme/gendarme-2.10-bin.z
ip

If I go to the download page:
https://github.com/spouliot/gendarme/downloads

The URL looks like it should now be:
https://github.com/downloads/spouliot/gendarme/gendarme-2.10-bin.zip

However, after this it still failed due to Github wanting to redirect.
I had to add a `-L` switch to the `curl` command to allow redirects.

I think we can close #220 in favor of this.