[generator] better support for "package-private" classes
Fixes https://github.com/xamarin/java.interop/issues/168

Java allows public classes to inherit from "package-private" classes, in
which case the `public` & `protected` members of the "package-private"
class are visible within the public class:

    // Java
    /* package */ abstract class SpannableStringInternal {
      public void getChars(int start, int end, char[] dest, int off);
      // ...
    }
    public class SpannableString extends SpannableStringInternal {
      // ...
    }

The problem is that `generator` didn't properly support this construct,
and *skips* binding of "package-private" types, resulting in generated
C# code such as:

    // C#
    public partial class SpannableString : SpannableStringInternal {
      // CS0246: The type or namespace name `SpannableStringInternal` could not be found
    }

This could be worked aroudn via Metadata by making the "package-private"
class `public`, but this means that if (when?) the "package-private"
class is *removed*, we have an ABI break.

Support this construct by updating `generator` to "copy" the members
from the "package-private" class into the declaring class:

    // C#
    public partial class SpannableString : Java.Lang.Object {
      public void GetChars(int start, int end, char[] dest, int off);
      // ...
    }

This allows the generated code to compile without metadata fixup.

Specifics in implementing this:
- Add a `FixupAccessModifiers` method to `GenBase`, later this may
need to be further extended for methods
- Call `FixupAccessModifiers` in the "validation" step
- Added a setter for `BaseType` so it can be modified
- In `FixupAccessModifiers`, lookup the base class of the current type
and check if it is non-public
- Skip the package-private types, and replace them with that class's
base type
- Look for each method on the base type, and copy it to the public type
if it does not exist
- Added tests for this scenario