method parameter type may contain "params " and cannot be used by fields.
This fixes https://github.com/xamarin/xamarin-android/issues/1138

It is quite corner case, where ... arguments are used in interfaces that
can be converted to events.

Parameter types cannot be simply used for fields and parameters because
they cannot take "params" as their modifiers.

The real code fix should be in Parameter.cs, but that would require
a lot of changes unlike this tiny fix.