[Java.Interop] Use ref return for elements access
While reviewing
https://github.com/spouliot/gendarme/wiki/Gendarme.Rules.Exceptions.DoNotThrowInUnexpectedLocationRule(2.10)
reported for JniArrayElements::Elements we discussed the issue with
Jonathan Pryor and came out with this API change to use `return ref`
instead of returning `IntPtr`. We also considered using
`Span<T>/Memory<T>`, it is not avaible in mono yet though.

The `Elements` property remains public, so that it can be used by 3rd
party code on primitive arrays. We also added `Size` property for that
reason.