[build] Generate $(JdkBinPath)` property
We're trying to build xamarin-android on a VSTS macOS instance which
has multiple Java JDKs installed, including JDK 9.

The problem is that the Android SDK and JDK 9 don't get along. The
`xamarin-android/tests/CodeGen-Binding/Xamarin.Android.LibraryProjectZip-LibBinding`
build invokes `gradlew`, `gradlew` uses whatever `java` is in `$PATH`,
and when using JDK 9 `gradlew` fails:

	Executing: ./gradlew assembleDebug --stacktrace
	...
	org.gradle.api.ProjectConfigurationException: A problem occurred configuring project ':library'.
	  ...
	Caused by: org.gradle.internal.event.ListenerNotificationException:
	Failed to notify project evaluation listener.
	  ...
	Caused by: java.lang.NoClassDefFoundError: javax/xml/bind/annotation/XmlSchema
	  ...
	Caused by: java.lang.ClassNotFoundException: javax.xml.bind.annotation.XmlSchema

(Rephrased: JDK 9 removed the type
`javax.xml.bind.annotation.XmlSchema`, which broke some dependency.)

55c56f7a was an attempt to support this environment, by *allowing*
JDK 9 to be filtered out when computing `$(JdkJvmPath)`,
`$(JavaCPath)`, and `$(JarPath)`.

Unfortunately this was inadequate: we need to override `$PATH` so that
the "right" JDK is used, not whatever happens to be first in `$PATH`.

In manual testing, this fails:

	ANDROID_HOME=... ./gradlew assembleDebug --stacktrace

while this works:

	ANDROID_HOME=... PATH=/path/to/jdk8/bin:$PATH ./gradlew assembleDebug --stacktrace --no-daemon

(The `--no-daemon` is needed so that we don't use some previously
launched gradle daemon under the wrong JDK.)

Thus, we want a way to easily insert the appropriate JDK path into
`$PATH` for the `gradlew` command.

Export a new `$(JdkBinPath)` MSBuild property which contains the JDK
`bin` directory. This will allow us to fix the
`Xamarin.Android.LibraryProjectZip-LibBinding` project:

	<Exec
	    EnvironmentVariables="ANDROID_HOME=$(AndroidSdkDirectory)"
	    Command="PATH=$(JdkBinPath):%24PATH .\gradlew assembleDebug --stacktrace --no-daemon"
	    WorkingDirectory="$(MSBuildThisFileDirectory)java\JavaLib"
	/>

and (hopefully) allow us to build on our VSTS bot which contains both
JDK 8 and 9.