Don't omit the extended API description attributes
class-parse emits a number of useful attributes which contain information about
the JNI signatures, return and parameter types etc. Currently, API XML adjuster
removes them from its output.
The information in extended attributes is (will be) useful for code generation
and thus this commit makes sure that all of them made it to the adjusted XML.