[Java.Interop] Fix Runtime's Dispose method
The dispose method of Runtime class wasn't working right when called
from finalizer.

That usually doesn't happen on Android, so it wasn't noticed until
now. It can happen even on Android though, when calling
System.Environment.Exit method.

In that case, the JniEnvironment's Info ThreadLocal object is already
disposed, so we cannot use it. It is also used in
JniObjectReference.Dispose so we should not try to call that as well.