[Java.Interop.Tools.Cecil] Change DirectoryAssemblyResolver to allow forced loading.
Context https://github.com/xamarin/xamarin-android/issues/1154

One of the problems we currently face is that our build system
resolves `ref` netstandard libraries. With the current cache
system, onces an assembly has been loaded into the Cecil cache
it only ever uses that assembly. We might need to replace the
current cached version with a new one if we detect a `ref` and
need to reload the `lib`.