[generator] fix for "this" keyword in parameter names
Context: https://devdiv.visualstudio.com/DevDiv/Xamarin%20Project%20Mgrs/_queries?id=568918

Apparently, it is possible for generator to receive the following method
as input:

    <method name="useThis" return="java.lang.String" static="true" visibility="public">
      <parameter name="this" type="java.lang.String"/>
    </method>

I wasn't able to create a method like this in Java (Android Studio), so
it must only be possible through some kind of Java bytecode
post-compilation step, or another JVM language. (perhaps Kotlin?)

The fix here was to modify `Parameter.ToNative` so it does not call
`CodeGenerationOptions.GetSafeIdentifier` for all cases. All
implementations of `ISymbol.Call` call `GetSafeIdentifier`, so I think
it was double-escaping the keyword in this example.

Other changes:

I found one of my older commits 968b474 was causing the
`CSharpKeywords.cs` expected file to not get copied to the correct
output directory. It needed to specify directories to go to the right
place. This meant that parts of the `CSharpKeywords` test wasn't being
verified--luckily it was all still working.