[tests] better assertion failure for Windows devs
I've been doing all my development on Windows lately (as a way to
experience what our Windows userbase encounters). One of the painpoints
of working on generator is the fact that the failing assertion message
is not very useful. I changed the message to a valid powershell command,
which seems inherently more useful on Windows.

This allows me to just paste error into powershell and view the diff
like macOS/linux users can do with the unix/bash command.