[tests] use wildcards for simpler generator integration tests
One of the pain points in setting up integration tests for generator is
adding new files to `expected.targets`. This is an easy step to forget,
and your new test won't actually be asserting against anything if the
files aren't copied to the output directory.

So a much better situation would be to:
- Use wildcards for everything under `expected\**\*` and
`expected.ji\**\*` to be copied to output
- Delete `expected.targets`
- New tests can just add files in the directory, that's it

When making this change, I invariably found some mistakes:
- Various tests had a `Java.Lang.Object.cs` file that was not needed/or
asserted against
- The `InterfaceMethodsConflict` test had some files that were not
asserted against, and needed the `new` keyword added for the latest
changes
- The `CSharpKeywords` tests had some differences in the output for
Windows vs macOS. It seemed better to just have the same output on both
platforms. A small change to check specifically for the `partial` keyword
accounts for the difference in the `CSharpCodeProvider.IsValidIdentifier`
method on macOS.

Other changes:
- Imported `RunNUnitTests.targets` into `Before.Java.Interop.sln.targets`
- This lets you do `msbuild Java.Interop.sln /t:RunTests`