[Java.Interop.Tools.Cecil] new exception type for DirectoryAssemblyResolver
Context: https://github.com/xamarin/xamarin-android/issues/1532

In discussion around this issue, we realized the error message is a bit
confusing for developers. After looking into it, the exception is thrown
inside java.interop, so it makes it a bit tough to do something specific
within an MSBuild task in xamarin-android.

By making a new `AssemblyNotFoundException` type, we can handle it
upstream in xamarin-android in the `ResolveAssemblies` MSBuild task, for
example. There we could have a specific `catch` statement and use the
`AssemblyNameReference` instance to present a better error message.

Currently users are getting:

    Could not load assembly XYZ. Perhaps it doesn't exist in the Mono for Android profile?

This message makes sense for `System.*` or framework assemblies.

We were thinking in some cases we could give an error such as:

    Could not load assembly XYZ. It was referenced by assembly ABC, but was not found. Are you missing a reference or nuget package?