Added AssemblyNotFoundException.ResolutionPath
To give further error detail, we need to record the list of assemblies
as `DirectoryAssemblyResolver` sees them.

To do this, I added a `resolutionPath` member variable which collects
assemblies as `IAssemblyResolver.Resolve` is called. This method gets
internally called by Mono.Cecil as well as directly in some places.
Usages of `DirectoryAssemblyResolver` are wrapped in `using` statements,
so a single list for the lifetime of the class should work here. We also
shouldn't add to the list if the assembly is already found in the cache.

I also added a `GetDetailedMessage` method so it could be used upstream
in xamarin-android. The `ResolveAssemblies` task will need to catch
`AssemblyNotFoundException` specifically and check
`MonoAndroidHelper.IsReferenceAssembly` to decide if we want either the
`Perhaps it doesn't exist in the Mono for Android profile?` or
`Are you missing a reference or nuget package?` message.

The relevant `ResolveAssemblies` source found here:

https://github.com/xamarin/xamarin-android/blob/aee709d2ec98ddeb9a0cec9604e4fefa7321196e/src/Xamarin.Android.Build.Tasks/Tasks/ResolveAssemblies.cs#L93-L127