[Java.Interop] Do not throw in RegisterNativeMembers
Do not throw the `NotSupportedException` exception in
`JniRuntime+JniTypeManager::RegisterNativeMembers` anymore. Throwing
that exception ment that every type to be registered had to have
register method to register marshaling methods, even when there were
no methods to register.

With recent changes by
https://github.com/xamarin/java.interop/commit/3f96359469b09602dc69bfa4b658a27c9128f243,
where we started to throw that exception again, we run into issues as
JavaProxyThrowable didn't have the *empty* register method
anymore. That led to a crash in the PropagateException tests, where
JavaProxyThrowable needed to be ready, but wasn't becase its
registration failed and thrown the exception, which cannot be
propagated because we didn't have JavaProxyThrowable in valid state.

To provide a way for XA's `AndroidRuntime.AndroidTypeManager` to know
whether the register method was found and used, we introduce new
protected `TryRegisterNativeMembers` method, which returns *bool*
value with that information, so that XA can do legacy fallback.

It also fixes Java side registration of JavaProxyThrowable, which was
using wrong class.