[Java.Interop] Do not throw in RegisterNativeMembers (#297)
Commit 3f963594 inadvertently broke the world: when running the
[`Java.Interop-Tess.dll` unit tests we SIGSEGV][0]:

[0]: https://jenkins.mono-project.com/view/Xamarin.Android/job/Java.Interop/223/console

	Executing: mono --debug packages/NUnit.ConsoleRunner.3.7.0/tools/nunit3-console.exe  bin/TestRelease/Java.Interop-Tests.dll  --result="TestResult-Java.Interop-Tests.xml;format=nunit2" --output="bin/TestRelease/TestOutput-Java.Interop-Tests.txt"
	...
	CallNonvirtualDerived.method() invoked!
	Stacktrace:
	  at <unknown> <0xffffffff>
	  at (wrapper managed-to-native) Java.Interop.NativeMethods.java_interop_jnienv_get_object_class (intptr,intptr)
	  at Java.Interop.JniEnvironment/Types.GetObjectClass (Java.Interop.JniObjectReference)
	  at Java.Interop.JniEnvironment/Types.GetJniTypeNameFromInstance (Java.Interop.JniObjectReference)
	  at Java.Interop.JniRuntime/JniValueManager.DisposePeer (Java.Interop.JniObjectReference,Java.Interop.IJavaPeerable)
	  at Java.Interop.JniRuntime/JniValueManager.DisposePeer (Java.Interop.IJavaPeerable)
	  at Java.Interop.JavaObject.Dispose ()
	  at Java.InteropTests.TestTypeTests.PropogateException ()
	  at (wrapper runtime-invoke) object.runtime_invoke_void__this__ (object,intptr,intptr,intptr)
	  ...
	Native stacktrace:
		0   mono                                0x000000010b58b251 mono_handle_native_crash + 257
		1   mono                                0x000000010b6279f6 altstack_handle_and_restore + 70
		2   libjvm.dylib                        0x00000001134bdab6 _ZN17HandleMarkCleanerD2Ev + 22
		3   libjvm.dylib                        0x00000001134ac6af jni_GetObjectClass + 179
		4   libjvm.dylib                        0x00000001134c0c11 checked_jni_GetObjectClass + 239
		5   libjava-interop.dylib               0x0000000129b55ad3 java_interop_jnienv_get_object_class + 51
		6   ???                                 0x000000012a08d395 0x0 + 5000188821
		7   ???                                 0x000000012a09df53 0x0 + 5000257363
		8   ???                                 0x0000000113143d59 0x0 + 4615060825
		9   mono                                0x000000010b4e03d7 mono_jit_runtime_invoke + 1383
		10  mono                                0x000000010b6d81f4 do_runtime_invoke + 84
		11  mono                                0x000000010b6dbd86 mono_runtime_try_invoke_array + 1254
		...

The problem was an unexpected interaction with 7d511631 and
"bad semantics" for
`Java.Interop.JniRuntime.JniTypeManager.RegisterNativeMembers()`:
`RegisterNativeMembers()` had *always* required that the type being
registered actually contain a "marshal method registration method",
and if there was no such method then a `NotSupportedException` would
be thrown; see c8f3e51a. (This "marshal method registration method"
was originally a `__RegisterNativeMembers()` static method declared
on the type being registered, and changed in 7d511631.)

The problems with this semantic are twofold:

 1. Not all types need to register native methods, yet this semantic
    requires that they contain registration methods anyway, and

 2. If a refactoring removes a "marshal method registration method",
    things start blowing up.

...which is what happened in 7d511631:
`JavaProxyThrowable.__RegisterNativeMembers()` was removed, and no
equivalent `JniAddNativeMethodRegistrationAttribute`-based
replacement was added. After 7d511631, attempts to register
`JavaProxyThrowable` would thus silently "fail," but because there
was no "error checking" for this failure, no exception was thrown,
and things kept working.

Then comes 3f963594, which *added the missing error checking*.
With 3f963594, registering `JavaProxyThrowable` would now cause a
`NotSupportedException` to be thrown, but because we need
`JavaProxyThrowable` in order to support marshaling exceptions from
managed to Java code, and registering `JavaProxyThrowable` failed,
everything fell to pieces.

(Aside: Jenkins, meanwhile, was misconfigured, preventing it from
reporting [the breakage within the PR build][1]), which is "green"
but contains the SIGSEGV mentioned above; it should not be green!
This has since been fixed.)

[1]: https://jenkins.mono-project.com/job/Java.Interop-pr-builder/399/

The primary fix is to change the semantics of
`RegisterNativeMembers()` so that it *does not throw*, ***ever***.
This removes the need for managed types to always provide a
"marshal method registration method", even if one isn't needed.
This in turn prevents the `NotSupportedException` from being thrown,
which allows things to behave as intended.

Additionally, `JavaProxyThrowable.java` had a "typo": it was
attempting to associate
`com.xamarin.java_interop.internal.JavaProxyObject` with
`Java.Interop.JavaProxyThrowable`, which is wrong; it should be
`com.xamarin.java_interop.internal.JavaProxyThrowable`.

Finally, add a new `protected` method:

	partial class JniRuntime {
	  partial class JniTypeManager {
	    protected bool TryRegisterNativeMembers (JniType nativeClass, Type type, string methods);
	  }
	}

This allows xamarin-android's `Android.Runtime.AndroidTypeManager` to
know  whether the register method was found and used, allowing it to
try to use the `jnimarshalmethod-gen.exe`-generated registration
*first*, then fallback to the legacy Reflection-based mechanism.