[generator] fix for NRE in `FixupAccessModifiers`
Context 4ec5d4e

In improving our support for "package-private" classes, I had to add a
check to compare if two methods have the same signature. The expression
was of the form:

    m.Name == baseMethod.Name && m.Parameters.JavaSignature == baseMethod.Parameters.JavaSignature

The problem with this is that the `JavaSignature` property can throw an
NRE if one of the parameter types is unknown, or only becomes valid
later on in generator's process. Instead I need to only rely on the
original Java information here, and loop over the parameters individually.

To fix this:
- Create a `bool Matches (MethodBase other)` method in `MethodBase`
- Override `Matches` in `Method`, also check the return type
- Add some unit tests for these scenarios for both `XmlMethod` and
`ManagedMethod`