[jnimarshalmethod-gen] Update the originating assembly
Update the originating assembly instead of creating a *new* one,
called `<AssemblyName>-new.dll`

I expected I would need to load the originating assembly in the
separate AppDomain, but instead it was enough to load it with cecil
with reading parameters having `ReadWrite = true`.

Because the referenced assemblies might not be writable, we need to
keep loading them as read only, thus with default `ReadWrite = false`.

For that I added new method to our cecil assembly resolver, to add
existing assembly definition to its cache. We first read the
originating assemblies as *ReadWrite* and add them to the resolver
cache.