[generator] SymbolTable is no longer static
While refactoring/writing tests for `generator` some issues arise
because `SymbolTable` is `static`.

At the begin and end of each test, we would like to begin with a clean
state. Unfortunately if `SymbolTable` is static, symbols are leftover
from any previous tests. This both introduces different behavior
depending on the ordering of tests, as well as potential interference
between tests.

Changes:
- `SymbolTable` is non-static, several helper methods that were
"stateless" were left to be static
- Added a property to `CodeGenerationOptions` to hold an instance of
`SymbolTable` since this object is already passed through most of
`generator`
- There were a few places I had to add additional parameters to pass the
instance of `CodeGenerationOptions` through
- The constructor for `ManagedGenBaseSupport` was hitting the
`SymbolTable`, but did not have a reference to `CodeGenerationOptions`,
so I moved this logic to an `OnValidate` method, renaming the
`ValidateNamespace` method. This seems to more closely match what is
happening in other classes in generator anyway.
- I was able to remove some initial hacks I added to the unit tests to
workaround `SymbolTable` being static
- I could also remove the `ISymbol managedSymbol` from `TestField`
completely.
- Since tests now start at a clean state, we can add to the
`SymbolTable` as needed during unit tests.
- I noticed that we weren't setting up `java.lang.Object` and
`java.lang.String` in both `XmlTests` and `ManagedTests`