[generator] fix regression for types from assemblies
Context: https://github.com/xamarin/java.interop/commit/19bcca50934ff7a2464e8d6170eadbba10c2b266

One of the side-effects of 19bcca5 was that `Validate` *has* to be
called in order for `ManagedClassGen` and `ManagedInterfaceGen` to be
initialized properly. Otherwise, properties such as `JavaName` are
going to return `null`.

I added a call to `Validate` in `CodeGenerator.ProcessReferencedType`,
and additionally added a new test class for the `Adjuster` class. For
now, it is verifying this regression is fixed, but should be expanded
down the road to get better test coverage around the `Adjuster` class.