[generator] fix regression for types from assemblies (#308)
Context: https://github.com/xamarin/xamarin-android/pull/1633

One of the side-effects of commit 19bcca5 was that now `Validate()`
*must* be called in order for `ManagedClassGen` and
`ManagedInterfaceGen` to be initialized properly.  Otherwise,
properties such as `JavaName` will return `null`, at which point
`generator --only-xml-adjuster` [no longer works][0]:

[0]: https://github.com/xamarin/xamarin-android/pull/1633#issuecomment-386339443

	$ generator.exe obj/Debug/api.xml.class-parse --only-xml-adjuster --xml-adjuster-output=obj/Debug/api.xml ...
	resolving assembly /Volumes/Seagate4TB/work/xamarin-android/bin/Debug/lib/xamarin.android/xbuild-frameworks/MonoAndroid/v1.0/mscorlib.dll. (TaskId:1143)
	resolving assembly /Volumes/Seagate4TB/work/xamarin-android/bin/Debug/lib/xamarin.android/xbuild-frameworks/MonoAndroid/v1.0/Java.Interop.dll. (TaskId:1143)
	...
	Error while processing type '[Class] com.xamarin.android.Bxc37706Throwable': Type 'java.lang.Throwable' was not found. (TaskId:1143)
	Error while processing type '[Class] com.xamarin.android.Bxc4288': Type 'java.lang.Object' was not found. (TaskId:1143)
	...

The result is that `.jar` bindings which previously worked, don't,
with lots of types missing from the adjusted `api.xml`.

Update `CodeGenerator.ProcessReferencedType()` to call `Validate()`,
and additionally added a new test class for the `Adjuster` class.
For now, it is verifying this regression is fixed, but should be
expanded down the road to get better test coverage around the
`Adjuster` class.