[build] Don't use ls(1) to find JDK paths (#314)
Fixes: https://github.com/xamarin/xamarin-android/issues/1493

On some machines, `make prepare` fails:

	build-tools/scripts/jdk.mk:130: *** missing separator.  Stop.
	make: *** [prepare-external] Error 2

Eventually, we "found" the "cause": This make fragment:

	$(shell ls -dtr $(_DARWIN_JDK_FALLBACK_DIRS) | sort | tail -1)

was dying a terrible horrible no good death:

	_DARWIN_JDK_ROOT=/Library/Java/JavaVirtualMachines/jdk1.8.0_161.jdk
	kkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkin/sh: 49m: command not found

[Turns Out™][0], **ls**(1) should be *avoided*, as its output is
*unsafe* (I'm not sure *why*, and I'm not able to repro the above
failure on my machine, but it's *clearly* bad).

[0]: http://mywiki.wooledge.org/ParsingLs

Why are we using **ls**(1)?  To sort by timestamp, via `ls -dtr`.
What's the recommended replacement?

> If you truly need a list of all the files in a directory in order
> by mtime so that you can process them in sequence, switch to perl,
> and have your perl program do its own directory opening and sorting

LOL?

Which brings us to the solution: we don't want to use Perl -- we
*want* something plausibly cross-platform -- so let's use our existing
cross-platform dependency: MSBuild!

Update the `<JdkInfo/>` task so that in addition to probing a variety
of Windows-specific registry locations, it can now do `jdk.mk`-style
directory probing, allowing us to find the maximum installed JDK
version which is (optionally) less than `$(JI_MAX_JDK)`.

Additionally, further enhance `<JdkInfo/>` so that it will *also*
check `$JAVA_HOME` and execute the following command:

	java -XshowSettings:properties -version 2>&1 | grep java.home

`$JAVA_HOME` is preferred, if specified and it fulfills the
requirements of `$(JI_MAX_JDK)`.

Additionally, drastically simplify `jdk.mk`.  Instead of computing the
desired JDK directory *on every build*, instead generate a new
`bin/Build$(CONFIGURATION)/JdkInfo.mk` file which contains the JDK
informatino.  The `make prepare` target will generate this file.

This approach simplifes `jdk.mk` -- trying to maintain it was
beginning to give me a headache, so as part of this we're dropping
support for 32-bit JVMs on macOS, as if anyone uses those -- and also
brings the macOS/Linux build closer-in-spirit to Windows, which was
already using the `<JdkInfo/>` task.