Now generator generates missing (IPeerable) casts so that code compiles.
CodeGenerator.WriteMethodBody() now takes CodeGen so that it can detect
InterfaceGen and if it is for interface then it adds missing casts.

It is also possible to simply add cast everywhere which makes code
generation simple, but it will involve ALL the unit tests' expected output
rewritten. It will be annoying work especially the branch development
is going to take a while until merged, so I skip that for now.