[generator] Fix and extend marking of "this is overriding DIM" step.
DIMs were not correctly marked as expected - they should be marking
only implementors of DIMs. What it was trying instead was just marking
anything that matches the signature, in the interfaces.

First this change expands the target of marking to class methods,
therefore anything that implements DIMs are covered.
Then limited the marking targets to only DIMs, not all the iface methods.
(The change result for expanding them to classes was disasterous
because of this problem.)