[generator] change DIM override marking.
For an DIM that overrides another method from its implementing interface,
it must be declared explicitly overriding the base interface method (which
can be either DIM or non-DIM).
To achieve that, we need to store the name of the declaring type of the
base method.

Also, we cannot "override" static methods, so skip checking that.