[build] remove XAIntegrationDebug and XAIntegrationRelease
In PR: https://github.com/xamarin/java.interop/pull/360

I will likely add a new `$(Configuration)` to be used in
xamarin-android, so we can build `Java.Interop` as a non-PCL.

The existence of `XAIntegrationDebug` and `XAIntegrationRelease` are
confusing, since they are no longer used.

See:
https://github.com/xamarin/java.interop/commit/54a2aff091193c0a64705b89fb6aab9d4359616e
https://github.com/xamarin/xamarin-android/pull/1231

This removes `XAIntegrationDebug` and `XAIntegrationRelease` from all
solutions, project files, and `Makefile`.

I also let VS 2017 update the SLN file, as it was an old VS 2010
format...