[JavaCallableWrappers] Parallel.ForEach per assembly
Since we are looking at every type in each assembly, we can
parallelize this work to make things a bit faster.

However, there are a few concerns when doing this:
- Made sure to add a *new* method `GetJavaTypesInParallel` so the
  caller can opt-in to this
- The `cache` in `DirectoryAssemblyResolver` needs to now be a
  `ConcurrentDictionary`
- The list of `javaTypes` returned by `JavaTypeScanner` needs to also
  use a `BlockingCollection`
- Downstream in `xamarin-android`, we need to make sure
  `GenerateJavaStubs` is now an `AsyncTask`, so the "thread safe"
  logger methods are used.

To time the improvements, I used this project in Xamarin.Forms:
https://github.com/xamarin/Xamarin.Forms/tree/master/Xamarin.Forms.ControlGallery.Android

Before I did anything, the times were:

    2594 ms  GenerateJavaStubs                          2 calls

After making things parallel:

    2143 ms  GenerateJavaStubs                          2 calls

And then after a few more LINQ improvements in the `GenerateJavaStubs`
MSBuild task:

    2104 ms  GenerateJavaStubs                          2 calls

So we have close to a 500ms improvement for `GenerateJavaStubs` in
this project, that will build a library project + application project.

To see the full changes required downstream in `xamarin-android`:
https://github.com/xamarin/xamarin-android/compare/master...jonathanpeppers:generatejavastubs-perf