[jcw-gen] Add TypeReference partial assembly name extension method
Context: https://github.com/xamarin/xamarin-android/pull/2349

This commit introduces a new extension method `GetPartialAssemblyName()`
that provides just the assembly name portion of
`GetPartialAssemblyQualifiedName()` so that `GenerateJavaStubs` doesn't
need to do any string splitting to get the assembly name for warning
XA4214.  It also makes a small replacement in `JavaNativeTypeManager` to
take advantage of the new extension method for readability.

This new extension method returns a partial assembly name that is
specifically just the *first* part of a fully qualified assembly name.
The documentation for [System.Reflection.AssemblyName][0] and
[`Type.GetType()`][1] refers to this first part as the *simple name* of
the assembly.

Another detail about terminology that might be useful to mention is that
a type name with no qualifiers is called the *simple name* of the type,
while a type name qualified with either a partial or full assembly name
is called *fully qualified*.  So for example, the
`GetPartialAssemblyQualifiedName()` extension method returns a fully
qualified type name, even though it is only qualified with a partial
assembly name.

[0]: https://docs.microsoft.com/dotnet/api/system.reflection.assemblyname?view=netframework-4.7.2#remarks
[1]: https://docs.microsoft.com/dotnet/api/system.type.gettype?view=netframework-4.7.2#remarks