Make logging calls slightly faster for disabled categories
Calls to `log_{info,debug}` may sometimes take parameters which will have to be
processed (functions called, memory allocated etc) even though the category the
calls use is disabled. This is because the category check is done inside the
`log_*` functions. This commit makes it slightly faster by wrapping the log call
in a macro which performs the check before calling the function, thus saving
some time. In Xamarin.Android native runtime this approach saved between 17 and
21ms (!) of startup time.