[jcwgen] add public method for GetDestinationPath
Context: https://github.com/xamarin/xamarin-android/pull/2505#issuecomment-445374019

In a world order where we are creating less temp files in
Xamarin.Android, it would be nice to leverage an overload of
`JavaCallableWrapperGenerator` that uses a `TextWriter` in memory.

Currently we are doing:

    //jti is a JavaCallableWrapperGenerator instance
    jti.Generate (outputPath);

    //Then later in our MSBuild task
    foreach (var file in Directory.GetFiles (temp, "*", SearchOption.AllDirectories)) {
        ...
        MonoAndroidHelper.CopyIfChanged (file, dest);
    }

So this is slower than it could be:

1. Writes to a bunch of temp files.
2. Recurses the directory where all these temp files exist.
3. Does a hash comparison of the contents of the source and
   destination.
4. Copies the file if needed.
5. Deletes all the temp files.

Instead, we could do something like:

    using (var memoryStream = new MemoryStream ())
    using (var writer = new StreamWriter (memoryStream)) {
        jti.Generate (writer);
        writer.Flush ();
        var path = jti.GetDestinationPath (outputPath);
        MonoAndroidHelper.CopyIfStreamChanged (memoryStream, path);
    }

This is quite a bit streamlined:

1. Generate the file in-memory.
2. Do a hash comparison of the in-memory file versus the destination.
3. Write to the file if needed.

The change we need in Java.Interop here is to add the
`GetDestinationPath` method. Further changes downstream in
xamarin-android will allow us reap the benefits.

I also improved the code a bit, no longer using `string.Split`. We can
do a simple `string.Replace` from `.` to `Path.DirectorySeparatorChar`,
and it will allocate less strings.